format view and update flow booking

+++ user_career_more/lib/expect_skills/controllers/skill_controller.dart
@@ -65,7 +65,7 @@ class SkillController extends AutoDisposeNotifier<SkillRequest>
     state =  state.copyWith(timeOnline: value);
   }
 
-  void setPriceOnline(double? value) {
+  void setPriceOnline(int? value) {
     state =  state.copyWith(priceOnline: value);
   }
 
@@ -73,7 +73,7 @@ class SkillController extends AutoDisposeNotifier<SkillRequest>
     state =  state.copyWith(timeOffline: value);
   }
 
-  void setPriceOffline(double? value) {
+  void setPriceOffline(int? value) {
     state =  state.copyWith(priceOffline: value);
   }
